Add a flash message that appears when login failed, move sendcode func to utils.go
