+++ main.go
@@ -2,35 +2,27 @@ package main
 
 import (
 	"context"
-	"crypto/tls"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
 
 	"github.com/MonopolyTechnic/simple-banking-system/models"
+	"github.com/flosch/pongo2/v4"
 	"github.com/gorilla/sessions"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"golang.org/x/crypto/bcrypt"
 
-	//"bytes"
-	"math/rand"
-	"net/smtp"
-	"time"
-
-	//"mime/multipart"
-	"net"
-	//"io"
 	"errors"
 )
 
 var (
 	host        string
 	port        string
-	env         map[string]string = readEnv(".env")
-	store                         = sessions.NewCookieStore([]byte("your-secret-key")) // Change this to a secure key
-	smsGateways                   = map[string]string{
+	env         map[string]string     = readEnv(".env")
+	store       *sessions.CookieStore = sessions.NewCookieStore([]byte("your-secret-key")) // Change this to a secure key
+	smsGateways map[string]string     = map[string]string{
 		"AT&T":               "txt.att.net",
 		"T-Mobile":           "tmomail.net",
 		"Verizon":            "vtext.com",
@@ -106,15 +98,19 @@ func index(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	http.ServeFile(w, r, "./templates/index.html")
+	RenderTemplate(w, "index.html")
 }
 
 func loginUser(w http.ResponseWriter, r *http.Request) {
-	http.ServeFile(w, r, "./templates/loginuser.html")
+	RenderTemplate(w, "loginuser.html", pongo2.Context{"flashes": RetrieveFlashes(r, w)})
 }
 
 func forgotPassword(w http.ResponseWriter, r *http.Request) {
-	http.ServeFile(w, r, "./templates/forgotpassword.html")
+	RenderTemplate(w, "forgotpassword.html")
+}
+
+func forgotPasswordSent(w http.ResponseWriter, r *http.Request) {
+	RenderTemplate(w, "postresetpassword.html")
 }
 
 func callback(w http.ResponseWriter, r *http.Request) {
@@ -147,9 +143,15 @@ func callback(w http.ResponseWriter, r *http.Request) {
 		return nil
 	})
 
+	session, err2 := store.Get(r, "session-name")
+	handle(err2)
+
 	// Invalid credentials
 	if err != nil {
-		// TODO: Show a message saying that invalid credentials have been entered
+		session.AddFlash("Invalid email or password entered.")
+		err2 = session.Save(r, w)
+		handle(err2)
+
 		http.Redirect(w, r, "/login-user", http.StatusSeeOther)
 		return
 	}
@@ -170,111 +172,8 @@ func verifyCode(w http.ResponseWriter, r *http.Request) {
 	http.Redirect(w, r, "/employee-dashboard", http.StatusSeeOther)
 }
 
-func forgotPasswordSent(w http.ResponseWriter, r *http.Request) {
-	http.ServeFile(w, r, "./templates/postresetpassword.html")
-}
-
 func employeeDashboard(w http.ResponseWriter, r *http.Request) {
-	http.ServeFile(w, r, "./templates/employeehomescreen.html")
-}
-
-func SendCode(phoneNumber, phoneCarrier string) (int, error) {
-	emailSender = env["EMAIL_SENDER"] // Read email sender from environment
-	emailPassword = env["EMAIL_PASSWORD"]
-
-	// Look up carrier gateway
-	carrierGateway, exists := smsGateways[phoneCarrier]
-	if !exists {
-		return 0, fmt.Errorf("unsupported carrier: %s", phoneCarrier)
-	}
-
-	recipientSMS := fmt.Sprintf("%s@%s", phoneNumber, carrierGateway)
-
-	// Generate a random 6-digit verification code
-	rand.Seed(time.Now().UnixNano())
-	verificationCode := rand.Intn(900000) + 100000
-
-	// Create the plain-text message body
-	message := fmt.Sprintf("Subject: Verification Code\n\nYour verification code is: %d", verificationCode)
-
-	// Set up SMTP connection
-	conn, err := net.Dial("tcp", "smtp.gmail.com:587")
-	if err != nil {
-		return 0, fmt.Errorf("could not connect to SMTP server: %v", err)
-	}
-
-	// Create SMTP client
-	host := "smtp.gmail.com"
-	c, err := smtp.NewClient(conn, host)
-	if err != nil {
-		return 0, fmt.Errorf("could not create SMTP client: %v", err)
-	}
-
-	// Upgrade to TLS
-	tlsConfig := &tls.Config{
-		ServerName: host,
-	}
-	if err := c.StartTLS(tlsConfig); err != nil {
-		return 0, fmt.Errorf("could not start TLS: %v", err)
-	}
-
-	// Authenticate using AUTH LOGIN
-	auth := LoginAuth(emailSender, emailPassword)
-	if err := c.Auth(auth); err != nil {
-		return 0, fmt.Errorf("could not authenticate: %v", err)
-	}
-
-	// Set the sender and recipient
-	if err := c.Mail(emailSender); err != nil {
-		return 0, fmt.Errorf("could not set sender: %v", err)
-	}
-	if err := c.Rcpt(recipientSMS); err != nil {
-		return 0, fmt.Errorf("could not set recipient: %v", err)
-	}
-
-	// Send the email
-	w, err := c.Data()
-	if err != nil {
-		return 0, fmt.Errorf("could not send data: %v", err)
-	}
-	if _, err := w.Write([]byte(message)); err != nil {
-		return 0, fmt.Errorf("could not write to SMTP: %v", err)
-	}
-	if err := w.Close(); err != nil {
-		return 0, fmt.Errorf("could not close SMTP connection: %v", err)
-	}
-
-	c.Quit()
-
-	// Return the verification code to the caller
-	return verificationCode, nil
-}
-
-// Define AUTH LOGIN
-type loginAuth struct {
-	username, password string
-}
-
-func LoginAuth(username, password string) smtp.Auth {
-	return &loginAuth{username, password}
-}
-
-func (a *loginAuth) Start(server *smtp.ServerInfo) (string, []byte, error) {
-	return "LOGIN", []byte(a.username), nil
-}
-
-func (a *loginAuth) Next(fromServer []byte, more bool) ([]byte, error) {
-	if more {
-		switch string(fromServer) {
-		case "Username:":
-			return []byte(a.username), nil
-		case "Password:":
-			return []byte(a.password), nil
-		default:
-			return nil, errors.New("unknown from server")
-		}
-	}
-	return nil, nil
+	RenderTemplate(w, "employeehomescreen.html")
 }
 
 func twofa(w http.ResponseWriter, r *http.Request) {
@@ -296,7 +195,7 @@ func twofa(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		http.ServeFile(w, r, "./templates/2fa.html")
+		RenderTemplate(w, "2fa.html")
 	} else if r.Method == http.MethodPost {
 		code := r.FormValue("code")
 		if actualCode, ok := session.Values["actualCode"].(string); ok && code == actualCode {
